Refactored and added banning events handlers for ban logging

+++ app/Models/BanModel.cs
@@ -6,7 +6,7 @@
         public string Name;
         public ulong ByUId;
         public string ByName;
-        public int ExpiresOn;
+        public long ExpiresOn;
         public string Expired;
         public string BannedOn;
         public string Reason;
